+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Changes before Tatum release are not documented in this file.
 
 #### Fixed
 
+- Fix operator review request event parsing (https://github.com/streamr-dev/network/pull/2714)
+
 #### Security